Rename symbol with type `ST_Other`.
Symbols that belongs to the BSS data section are flagged as `ST_Other`.

The following C code produces such a symbol:

```c
__thread int thread_local_var;
```

As shown by `nm`, `thread_local_var` belongs to the BSS data section:

```
$ nm a.o
0000000000000000 B thread_local_var
```

This commit adds the support of these symbols.

Additionally, a unit test for that case was added.